+++ R/RcppExports.R
@@ -39,7 +39,7 @@ marginalKFWER <- function(thr, Z) {
     .Call('sansSouci_marginalKFWER', PACKAGE = 'sansSouci', thr, Z)
 }
 
-#' partial sorting of the rows of a matrix by descending order
+#' partial sorting of the columns of a matrix by descending order
 #'
 #' @param X A numeric matrix
 #' @param k An integer value between 1 and \code{nrow(mat)}