+++ R/empiricalCoverage.R
@@ -1,5 +1,7 @@
 ##' @export
-empiricalCoverage <- function(thr, mat, kMax=nrow(mat)) {
+empiricalCoverage <- function(thr, mat) {
+    kMax <- length(thr)
+    stopifnot(kMax<=nrow(mat)) ## sanity check
     kmaxH0 <- partialColSortDesc(mat, kMax); ## Implicitly forcing 'Rcpp' flavor
     empiricalCoverageO(thr, kmaxH0);
 }